fix: check for local storage availability (#540)
`localStorage` can be disabled in some browsers and contexts and it
causes the bug screen.  It's one of the few remaining critical errors.
We need to check for local storage availability before using it.  In
cases where it's not available, we just won't store drafts.

Refs: #525

+++ web-application/client/pages/authentication/AgeGate.js
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
- *  Communities -- Non-profit, cooperative social media 
- *  Copyright (C) 2022 - 2024 Daniel Bingham 
+ *  Communities -- Non-profit, cooperative social media
+ *  Copyright (C) 2022 - 2024 Daniel Bingham
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as published
@@ -25,6 +25,7 @@ import logger from '/logger'
 import { deleteUser } from '/state/User'
 import { reset } from '/state/system'
 
+import { isLocalStorageAvailable } from '/lib/localStorage'
 import { useRequest } from '/lib/hooks/useRequest'
 
 import CommunitiesLogo from '/components/header/CommunitiesLogo'
@@ -42,9 +43,16 @@ const AgeGate = function() {
     useEffect(function() {
         if ( request && request.state == 'fulfilled') {
 
-            // Clear local storage so their drafts don't carry over to another
-            // login session.
-            localStorage.clear()
+
+            if ( isLocalStorageAvailable() ) {
+                try {
+                    // Clear local storage so their drafts don't carry over to another
+                    // login session.
+                    localStorage.clear()
+                } catch (error) {
+                    logger.error(error)
+                }
+            }
 
             dispatch(reset())
